Refactor app review scraper and utility functions for improved structure and readability
